Time limit checked - it works
d is the date object, and is called within a function, that's why you can't see the current value of d. What you see is a d object defined outside of the function - when the experiment starts, to get exp_start_time. That's why d and exp_start_time seem to be the same to you - they are the same :)